requirements.txt
functions-framework
google-cloud-storage
pytotp
websocket-client
NorenRestApiPy
pyotp
selenium